Move the parsing of Value in a separate function.
