+++ linera-execution/src/test_utils/solidity.rs
@@ -122,3 +122,21 @@ pub fn get_contract_service_paths(module: Vec<u8>) -> anyhow::Result<(PathBuf, P
     let evm_service = app_path.to_path_buf();
     Ok((evm_contract, evm_service, dir))
 }
+
+fn value_to_vec_u8(value: Value) -> Vec<u8> {
+    let mut vec: Vec<u8> = Vec::new();
+    for val in value.as_array().unwrap() {
+        let val = val.as_u64().unwrap();
+        let val = val as u8;
+        vec.push(val);
+    }
+    vec
+}
+
+fn read_evm_u64_entry(value: Value) -> anyhow::Result<u64> {
+    let vec = value_to_vec_u8(value);
+    let mut arr = [0_u8; 8];
+    arr.copy_from_slice(&vec[24..]);
+    let counter_value = u64::from_be_bytes(arr);
+    Ok(counter_value)
+}
